+++ src/Definitions/API.php
@@ -19,6 +19,13 @@ final class API
     public const V2 = 'v2';
 
     /**
+     * API version 3
+     *
+     * @var string
+     */
+    public const V3 = 'v3';
+
+    /**
      * API latest version
      *
      * @var string
@@ -35,5 +42,6 @@ final class API
     public const URL = [
         self::V1 => 'https://api.balikobot.cz/',
         self::V2 => 'https://api.balikobot.cz/v2/',
+        self::V3 => 'https://api.balikobot.cz/v3/',
     ];
 }