+++ tests/Unit/Balikobot/TrackPackageTest.php
@@ -15,9 +15,12 @@ class TrackPackageTest extends AbstractBalikobotTestCase
             'status' => 200,
             0        => [
                 [
-                    'date'      => '2018-11-07 14:15:01',
-                    'name'      => 'Přijetí',
-                    'status_id' => 2,
+                    'date'           => '2018-11-07 14:15:01',
+                    'name'           => 'Doručování zásilky',
+                    'status_id'      => 2,
+                    'status_id_v2'   => 2.2,
+                    'type'           => 'event',
+                    'name_balikobot' => 'Zásilka je v přepravě.',
                 ],
             ],
         ]);
@@ -31,7 +34,7 @@ class TrackPackageTest extends AbstractBalikobotTestCase
         $requester->shouldHaveReceived(
             'request',
             [
-                'https://api.balikobot.cz/v2/ppl/track',
+                'https://api.balikobot.cz/v3/ppl/track',
                 [
                     0 => [
                         'id' => '1234',
@@ -49,19 +52,28 @@ class TrackPackageTest extends AbstractBalikobotTestCase
             'status' => 200,
             0        => [
                 [
-                    'date'      => '2018-11-07 14:15:01',
-                    'name'      => 'Přijetí',
-                    'status_id' => 2,
+                    'date'           => '2018-11-07 14:15:01',
+                    'name'           => 'Doručování zásilky',
+                    'status_id'      => 2,
+                    'status_id_v2'   => 2.2,
+                    'type'           => 'event',
+                    'name_balikobot' => 'Zásilka je v přepravě.',
                 ],
                 [
-                    'date'      => '2018-11-08 18:00:00',
-                    'name'      => 'Doručení',
-                    'status_id' => 1,
+                    'date'           => '2018-11-08 18:00:00',
+                    'name'           => 'Dodání zásilky. (77072 - Depo Olomouc 72)',
+                    'status_id'      => 1,
+                    'status_id_v2'   => 1.2,
+                    'type'           => 'notification',
+                    'name_balikobot' => 'Zásilka byla doručena příjemci.',
                 ],
                 [
-                    'date'      => '2018-11-09 11:19:51',
-                    'name'      => 'Dodání',
-                    'status_id' => -1,
+                    'date'           => '2020-11-06 21=>00=>00',
+                    'name'           => 'Obdrženy údaje k zásilce.',
+                    'status_id'      => -1,
+                    'status_id_v2'   => -1,
+                    'type'           => 'event',
+                    'name_balikobot' => 'Zásilka zatím nebyla předána dopravci.',
                 ],
             ],
         ]);
@@ -73,10 +85,13 @@ class TrackPackageTest extends AbstractBalikobotTestCase
         $statuses = $service->trackPackage($package);
 
         $this->assertEquals(3, count($statuses));
-        $this->assertEquals(2, $statuses[0]->getId());
+        $this->assertEquals(2, $statuses[0]->getGroupId());
+        $this->assertEquals(2.2, $statuses[0]->getId());
         $this->assertEquals(new DateTime('2018-11-08 18:00:00'), $statuses[1]->getDate());
-        $this->assertEquals('Dodání', $statuses[2]->getName());
-        $this->assertEquals(1, $statuses[1]->getId());
+        $this->assertEquals('Obdrženy údaje k zásilce.', $statuses[2]->getDescription());
+        $this->assertEquals('Zásilka zatím nebyla předána dopravci.', $statuses[2]->getName());
+        $this->assertEquals('notification', $statuses[1]->getType());
+        $this->assertEquals(1.2, $statuses[1]->getId());
     }
 
     public function testMakeRequestWithMultiplePackages()
@@ -85,16 +100,22 @@ class TrackPackageTest extends AbstractBalikobotTestCase
             'status' => 200,
             0        => [
                 [
-                    'date'      => '2018-11-07 14:15:01',
-                    'name'      => 'Přijetí',
-                    'status_id' => 2,
+                    'date'           => '2018-11-07 14:15:01',
+                    'name'           => 'Doručování zásilky',
+                    'status_id'      => 2,
+                    'status_id_v2'   => 2.2,
+                    'type'           => 'event',
+                    'name_balikobot' => 'Zásilka je v přepravě.',
                 ],
             ],
             1        => [
                 [
-                    'date'      => '2018-11-08 14:18:06',
-                    'name'      => 'Odeslání',
-                    'status_id' => 3,
+                    'date'           => '2018-11-08 18:00:00',
+                    'name'           => 'Dodání zásilky. (77072 - Depo Olomouc 72)',
+                    'status_id'      => 1,
+                    'status_id_v2'   => 1.2,
+                    'type'           => 'event',
+                    'name_balikobot' => 'Zásilka byla doručena příjemci.',
                 ],
             ],
         ]);
@@ -110,7 +131,7 @@ class TrackPackageTest extends AbstractBalikobotTestCase
         $requester->shouldHaveReceived(
             'request',
             [
-                'https://api.balikobot.cz/v2/ppl/track',
+                'https://api.balikobot.cz/v3/ppl/track',
                 [
                     0 => [
                         'id' => '1236',
@@ -131,26 +152,38 @@ class TrackPackageTest extends AbstractBalikobotTestCase
             'status' => 200,
             0        => [
                 [
-                    'date'      => '2018-11-07 14:15:01',
-                    'name'      => 'Přijetí',
-                    'status_id' => 2,
+                    'date'           => '2018-11-07 14:15:01',
+                    'name'           => 'Doručování zásilky',
+                    'status_id'      => 2,
+                    'status_id_v2'   => 2.2,
+                    'type'           => 'event',
+                    'name_balikobot' => 'Zásilka je v přepravě.',
                 ],
                 [
-                    'date'      => '2018-11-08 18:00:00',
-                    'name'      => 'Doručení',
-                    'status_id' => 1,
+                    'date'           => '2018-11-08 18:00:00',
+                    'name'           => 'Dodání zásilky. (77072 - Depo Olomouc 72)',
+                    'status_id'      => 1,
+                    'status_id_v2'   => 1.2,
+                    'type'           => 'notification',
+                    'name_balikobot' => 'Zásilka byla doručena příjemci.',
                 ],
                 [
-                    'date'      => '2018-11-09 11:19:51',
-                    'name'      => 'Dodání',
-                    'status_id' => -1,
+                    'date'           => '2020-11-06 21=>00=>00',
+                    'name'           => 'Obdrženy údaje k zásilce.',
+                    'status_id'      => -1,
+                    'status_id_v2'   => -1,
+                    'type'           => 'event',
+                    'name_balikobot' => 'Zásilka zatím nebyla předána dopravci.',
                 ],
             ],
             1        => [
                 [
-                    'date'      => '2018-11-08 14:18:06',
-                    'name'      => 'Přijetí',
-                    'status_id' => 2,
+                    'date'           => '2018-11-08 14:18:06',
+                    'name'           => 'Dodání zásilky. (77072 - Depo Olomouc 72)',
+                    'status_id'      => 1,
+                    'status_id_v2'   => 1.1,
+                    'type'           => 'event',
+                    'name_balikobot' => 'Zásilka byla doručena příjemci.',
                 ],
             ],
         ]);
@@ -165,11 +198,13 @@ class TrackPackageTest extends AbstractBalikobotTestCase
 
         $this->assertCount(2, $statuses);
         $this->assertCount(3, $statuses[0]);
-        $this->assertEquals(2, $statuses[0][0]->getId());
+        $this->assertEquals(2, $statuses[0][0]->getGroupId());
+        $this->assertEquals(2.2, $statuses[0][0]->getId());
         $this->assertEquals(new DateTime('2018-11-08 18:00:00'), $statuses[0][1]->getDate());
-        $this->assertEquals('Dodání', $statuses[0][2]->getName());
-        $this->assertEquals(1, $statuses[0][1]->getId());
-        $this->assertEquals(2, $statuses[1][0]->getId());
+        $this->assertEquals('Obdrženy údaje k zásilce.', $statuses[0][2]->getDescription());
+        $this->assertEquals('Zásilka zatím nebyla předána dopravci.', $statuses[0][2]->getName());
+        $this->assertEquals(1.2, $statuses[0][1]->getId());
+        $this->assertEquals(1.1, $statuses[1][0]->getId());
         $this->assertEquals(new DateTime('2018-11-08 14:18:06'), $statuses[1][0]->getDate());
     }
 }